+++ CHANGELOG.md
@@ -17,6 +17,7 @@
  * **DSL change**: instance-based events inspection now returns event instances (instead of the event names as symbol)
  * **DSL change**: instance-based permissible_events has been removed in favor or events(:permissible => true)
  * **DSL change**: class-based events now returns a list of Event instances (instead of a hash with event names as keys)
+ * **DSL change**: renamed permissible states and events to permitted states events
 
 ## 3.9.0 (not yet released)